+++ src/papple2/Hooks.py
@@ -1,7 +1,14 @@
 #!/usr/bin/env python3
 
-from papple2.util import *
-from papple2.Emulator import *
+from papple2.util import hexaddr, hexbyte, lerp_rgb, chunks
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from papple2.Emulator import Emulator
+
+from papple2.Apple import Apple2
+from papple2.CPU import CPU
 import math