+++ tests/test_assembler.py
@@ -1,6 +1,8 @@
 import unittest
+from papple2.util import hexaddr
 from papple2.Assembler import *
-from papple2.Apple import *
+from papple2.CPU import CPU
+from papple2.Apple import Apple2
 
 
 class TestAssembler(unittest.TestCase):